requirements.txt
numpy
tensorflow
ray
cython
tqdm
tflite_runtime-2.6.0-cp37-cp37m-linux_x86_64.whl